Create test harness
Add a simple test harness for running layout tests.

Currently, the tests are run manually by opening test pages in the browser and checking the console for PASSED or FAILED reports.

Initial tests exercise the following animation element types: animate, animateTransform and set.

We create an animation using the Polyfill implementation, and another
using the native implementation, and verify property expectations at a
sequence of times.