Simple animation tests
Add a test that xlink:href can be used to specify the animated element.

Also add several other animation tests from https://github.com/ericwilligers/svg-animation.

We create an animation using the Polyfill implementation, and another using the native implementation, and verify property expectations at a sequence of times.

Unfortunately, the animated value of the fill property isn't being reported for the natively animated element, only for the element animated using the SMIL Polyfill. We animate the fill value in color.html and attempt to test in color-check.js (With the other properties being animated, this problem isn't occurring.)