Allow for mpath element without animateMotion parent
We were previously setting the mpathRecord member of undefined, when we assigned to
  animationRecords[element.parentNode].mpathRecord
without checking that element.parentNode was present in animationRecords.

We were assuming that the parent element of an mpath element would be an animateMotion element, and therefore already in animationRecords. This is the only context in which mpath is useful, but we shouldn't make assumptions about the SVG we will encounter.