getSimpleDuration and friends
Add the following to AnimationElement prototype:-
- targetElement property
- getSimpleDuration()
- getCurrentTime()
- getStartTime()

These are defined on SVG animation elements, even if the onload event
has not fired yet.

All duration and time values returned are in seconds.

To support testing of getCurrentTime, we modify the Web Animations polyfill so that, after _pauseAnimationsForTesting has been called, timeline currentTime is the pauseAt time.