Automated test for lists of begin, end times
We were calling _pauseAnimationsForTesting but that changes the players - in particular, it changes their start time. We now fix the animation clock at a particular time, without modifying the players.

When an animations ends, we were not handling the 'freeze' attribute correctly - we were letting every animation freeze at the values from the time the animation ended.