Support paced calcMode
Support the calcMode="paced" timing option. When this option is used, the timing for our polyfill animation matches that of the native animation, for motion path animation over a Bezier path. (With calcMode="linear", this does not occur as Chrome's native animation does not correctly support calcMode="linear".)

Automate test/testcases/animateMotion-autoReverse. This test wasn't previously automated as the polyfill and native implementation did not match in observed behavior.