Improve robustness of begin/end events test
If the machine is busy while running the begin-end-events test, the events may be issued before the test is ready for them. We move the begin and end times back tens of seconds to avoid such issues.

(This does not cause the test to run any more slowly, as we advance the clock once the test has started.)

This robustness change is important when we run many tests concurrently.