Respond to keystrokes
"Accesskey values allow an author to tie a begin or end time to a particular key press, independent of focus issues."
http://www.w3.org/TR/SMIL3/smil-timing.html#q29

begin="accessKey(1)" means the element begins when '1' is pressed.

We register a single listener on the document body when we first encounter a begin or end spec that uses accessKey.