Allow space or comma separated transform values
Scale and translate and rotate values may be specified with spaces or commas.

We convert to using commas before forwarding to Web Animations, with one
exception to be addressed in a subsequent CL:-

Web Animations support for rotate with spaces (not commas) was added in an earlier CL. We use that for now, but it should have used commas instead.