Support to-animation where from value is implicitly the current value
to animation is described in
http://www.w3.org/TR/2001/REC-smil-animation-20010904/#AnimFuncValues
"the animation function is defined to start with the underlying value for the attribute, and finish with the value specified with the to attribute. Using this form, an author can describe an animation that will start with any current value for the attribute, and will end up at the desired to value."

Also correct test animateMotion-paced test to use correct element names.