Hyphens and periods in ids can be escaped
The begin and end attributes of SMIL animation elements can refer to other elements by id. They can also contain hyphens and periods with specific meanings: a hyphen for negative offsets, and a period before an event name.

When an element id contains a hyphen or period, that character must be
escaped using a backslash.

http://www.w3.org/TR/SMIL3/smil-timing.html#q21